pyproject.toml
[tool.poetry]
name = "qa-guru-diploma"
version = "0.1.0"
description = ""
authors = ["marcelinleen <94484224+marcelinleen@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "qa_guru_diploma"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.4.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
selene = "2.0.0b5"
allure-pytest = "^2.13.2"
jsonschema = "^4.18.4"
appium-python-client = "2.11.1"
pydantic = "^2.1.1"
pydantic-settings = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
